+++ include/split_u.hpp
@@ -9,6 +9,6 @@ std::pair<
 >
 SplitU(
     const std::vector<Item>& U,
-    int threshold,
+    double threshold,
     size_t d
 );
